refactor(docs): extract REST API reference into lib/api-reference.ts
Third cut of the docs generator split. Moves the OpenAPI/curl reference
cluster into lib/api-reference.ts: the OpenApiOperation type, tagOrder,
isGenericObjectSchema, sampleOpenApiValue, firstOpenApiContent,
curlForOperation, renderApiReference. Pure formatting over the typed route
operation objects; the orchestrator (which builds operations via
packages/core/src/openapi) imports back OpenApiOperation + curlForOperation +
renderApiReference. Imports escapeHtml + generatedPageNote from doc-utils.

Pure relocation — bodies byte-for-byte unchanged. generate.ts 2,022 -> 1,925.
Verified: eslint (changed files) clean + docs:generate output BYTE-IDENTICAL
to baseline.
